+++ src/Directory.ts
@@ -12,7 +12,7 @@ class Directory {
         this.root = root || '/';
     }
 
-    walk(depth = 1) {
+    walk(depth = 1, verbose = false) {
         const dontNeedToWalk = depth <= 0 || (this.children && this.depth >= depth);
         if (dontNeedToWalk) {
             return this;
@@ -20,10 +20,13 @@ class Directory {
         try {
             this.depth = depth;
             const dirs = d(this.root);
+            if (verbose) {
+                console.log(`${this.root}: ${dirs}`);
+            }
             this.children = dirs.map((d: any) => new Directory(join(this.root, d)));
             if (--depth) {
                 try {
-                    this.children.forEach((c: any) => c.walk(depth));
+                    this.children.forEach((c: any) => c.walk(depth, verbose));
                 } catch (error) {
                     this.errors.push(error.message);
                 }